对于NUMBER(A,B) 形式的变量 把Integer修改为Double

+++ src/main/java/main/java/conversionRequest/BlazApCu5YearsOverdue.java
@@ -12,7 +12,7 @@ public class BlazApCu5YearsOverdue {
     private String F_ID;
     private String OVERDUE_MONTH;
     private Integer OVERDUE_MONTHS;
-    private Integer OVERDUE_AMOUNT;
+    private Double OVERDUE_AMOUNT;
     private String REPORT_ID;
 
     public String getP_ID() {
@@ -47,11 +47,11 @@ public class BlazApCu5YearsOverdue {
         this.OVERDUE_MONTHS = OVERDUE_MONTHS;
     }
 
-    public Integer getOVERDUE_AMOUNT() {
+    public Double getOVERDUE_AMOUNT() {
         return OVERDUE_AMOUNT;
     }
 
-    public void setOVERDUE_AMOUNT(Integer OVERDUE_AMOUNT) {
+    public void setOVERDUE_AMOUNT(Double OVERDUE_AMOUNT) {
         this.OVERDUE_AMOUNT = OVERDUE_AMOUNT;
     }
