+++ src/main/java/main/java/conversionRequest/BlazApHsAcctChgs.java
@@ -18,10 +18,10 @@ public class BlazApHsAcctChgs {
     private Integer LIMIT_CHWAY;
     private Integer LIMIT_T;
     private Integer CHGA_DATE;
-    private Integer CHGA_AFTER_VALUE;
+    private Double CHGA_AFTER_VALUE;
     private Integer CHGA_BEG_DT;
     private Integer CHGA_END_DT;
-    private Integer CHGA_BEFOR_VALUE;
+    private Double CHGA_BEFOR_VALUE;
     private Date MYETL_DATE;
 
     public String getCUSTR_NBR() {
@@ -112,11 +112,11 @@ public class BlazApHsAcctChgs {
         this.CHGA_DATE = CHGA_DATE;
     }
 
-    public Integer getCHGA_AFTER_VALUE() {
+    public Double getCHGA_AFTER_VALUE() {
         return CHGA_AFTER_VALUE;
     }
 
-    public void setCHGA_AFTER_VALUE(Integer CHGA_AFTER_VALUE) {
+    public void setCHGA_AFTER_VALUE(Double CHGA_AFTER_VALUE) {
         this.CHGA_AFTER_VALUE = CHGA_AFTER_VALUE;
     }
 
@@ -136,11 +136,11 @@ public class BlazApHsAcctChgs {
         this.CHGA_END_DT = CHGA_END_DT;
     }
 
-    public Integer getCHGA_BEFOR_VALUE() {
+    public Double getCHGA_BEFOR_VALUE() {
         return CHGA_BEFOR_VALUE;
     }
 
-    public void setCHGA_BEFOR_VALUE(Integer CHGA_BEFOR_VALUE) {
+    public void setCHGA_BEFOR_VALUE(Double CHGA_BEFOR_VALUE) {
         this.CHGA_BEFOR_VALUE = CHGA_BEFOR_VALUE;
     }
